removed Author entity
