--- config/validator/validation_author.yaml
@@ -1,8 +1,0 @@
-App\Entity\Author:
-    properties:
-        name:
-            - NotNull: ~
-            - Length:
-                max: 64
-            - Type:
-                type: string
