--- templates/author/edit.html.twig
@@ -1,10 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}
-    Edit author
-{% endblock %}
-{% block content %}
-    <h1 class="text-center">Edit author</h1>
-    {{ form_start(form, {'attr': {'novalidate': 'novalidate', 'class':'author-form'}}) }}
-    {{ form_widget(form) }}
-    {{ form_end(form) }}
-{% endblock %}
