--- templates/author/index.html.twig
@@ -1,28 +1,0 @@
-{% extends 'admin-index.html.twig' %}
-{% block title %}
-    Authors
-{% endblock %}
-{% block contentName %}
-    Authors
-    <a href="{{- path('author_new') -}}" class="btn btn-secondary btn-sm active" role="button">new</a>
-{% endblock %}
-
-{% block items %}
-    {% if authors|length == 0 %}
-    <p class="text-center">Authors not found.</p>
-    {% else %}
-    {% for author in authors %}
-        <tr>
-            <td>{{ loop.index }}</td>
-            <td>{{ author.name }}</td>
-            <td>
-                <a href="{{- path('author_edit', {id: author.id}) -}}" class="btn btn-secondary btn-sm active" role="button">edit</a>
-            </td>
-        <tr>
-    {% endfor %}
-    {% endif %}
-{% endblock %}
-
-{% block pagination %}
-    {{ knp_pagination_render(authors) }}
-{% endblock %}
