--- templates/author/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}
-  Create author
-{% endblock %}
-{% block content %}
-  <h1 class="text-center">Create author</h1>
-  {{ parent() }}
-  {{ form_start(form, {'attr': {'novalidate': 'novalidate', 'class':'author-form'}}) }}
-  {{ form_widget(form) }}
-  {{ form_end(form) }}
-{% endblock %}
