+++ templates/post/_post-item.html.twig
@@ -12,7 +12,7 @@
         <a href="{{- path('post_delete', {slug: post.slug}) -}}" class="btn btn-secondary btn-sm" role="button">Delete</a>
     </div>
     <div class="card-footer text-muted">
-        Posted on {{ post.createdAt|date('Y-m-d') }} by <a href="#">{{ post.author.name }}</a> <br>
+        Posted on {{ post.createdAt|date('Y-m-d') }} by <a href="#">{{ post.author.username }}</a> <br>
         Category: {{ post.category.title }} <br>
         Tags:
           {% for tag in post.tags %}
